# postman/postman-collection-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/orders/user/{userId}": {
            "get": {
                "description": "Retrieve orders for a specific user. Users can only access their own orders unless they are admin/moderator.",
                "operationId": "findByUser",
                "parameters": [{ "name": "userId", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot access other user's orders" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Get orders by user",
                "tags": ["Orders"]
            }
        },
...
        "/reviews": {
            "post": {
                "description": "Creates a new product review with rating and optional comment. Users can only create reviews for themselves.",
                "operationId": "create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "Review data including product_id, rating, and optional title/comment",
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateReviewDto" } } }
                },
                "responses": {
                    "201": {
                        "description": "Review successfully created",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "user_id": 1,
                                    "product_id": 1,
                                    "rating": 5,
                                    "title": "Great product!",
                                    "comment": "This product exceeded my expectations.",
                                    "is_verified_purchase": true,
                                    "helpful_votes": 0,
                                    "created_at": "2024-01-01T00:00:00Z",
                                    "updated_at": "2024-01-01T00:00:00Z"
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid input data" },
                    "401": { "description": "Unauthorized - Invalid or missing token" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Create a new review",
                "tags": ["Reviews"]
            },
            "get": {
                "description": "Retrieve reviews with support for pagination, filtering, sorting, and searching",
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": { "minimum": 1, "default": 1, "example": 1, "type": "number" }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": { "minimum": 1, "maximum": 100, "default": 10, "example": 10, "type": "number" }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "Number of records to skip (overrides page calculation)",
                        "schema": { "minimum": 0, "example": 0, "type": "number" }
                    },
                    {
                        "name": "sortBy",
                        "required": false,
                        "in": "query",
                        "description": "Field to sort by",
                        "schema": { "example": "created_at", "type": "string" }
                    },
                    {
                        "name": "sortOrder",
                        "required": false,
                        "in": "query",
                        "description": "Sort order",
                        "schema": { "default": "ASC", "example": "ASC", "type": "string", "enum": ["ASC", "DESC"] }
                    },
                    {
                        "name": "search",
                        "required": false,
                        "in": "query",
                        "description": "Search term to filter results",
                        "schema": { "example": "laptop", "type": "string" }
                    },
                    {
                        "name": "include",
                        "required": false,
                        "in": "query",
                        "description": "Related entities to include in the response (joins). Can be comma-separated string or array",
                        "schema": { "example": ["user", "product"], "type": "array", "items": { "type": "array" } }
                    },
                    {
                        "name": "user_id",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews by user ID",
                        "schema": { "example": 1, "type": "number" }
                    },
                    {
                        "name": "product_id",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews by product ID",
                        "schema": { "example": 1, "type": "number" }
                    },
                    {
                        "name": "rating",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews by exact rating",
                        "schema": { "minimum": 1, "maximum": 5, "example": 5, "type": "number" }
                    },
                    {
                        "name": "min_rating",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews with rating greater than or equal to this value",
                        "schema": { "minimum": 1, "maximum": 5, "example": 3, "type": "number" }
                    },
                    {
                        "name": "max_rating",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews with rating less than or equal to this value",
                        "schema": { "minimum": 1, "maximum": 5, "example": 4, "type": "number" }
                    },
                    {
                        "name": "is_verified_purchase",
                        "required": false,
                        "in": "query",
                        "description": "Filter by verified purchase status",
                        "schema": { "example": true, "type": "boolean" }
                    },
                    {
                        "name": "created_after",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews created after this date (ISO 8601 format)",
                        "schema": { "example": "2024-01-01T00:00:00Z", "type": "string" }
                    },
                    {
                        "name": "created_before",
                        "required": false,
                        "in": "query",
                        "description": "Filter reviews created before this date (ISO 8601 format)",
                        "schema": { "example": "2024-12-31T23:59:59Z", "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "user_id": 1,
                                            "product_id": 1,
                                            "rating": 5,
                                            "title": "Great product!",
                                            "comment": "This product exceeded my expectations.",
                                            "is_verified_purchase": true,
                                            "helpful_votes": 5,
                                            "created_at": "2024-01-01T00:00:00Z",
                                            "updated_at": "2024-01-01T00:00:00Z",
                                            "user": { "id": 1, "first_name": "John", "last_name": "Doe" },
                                            "product": { "id": 1, "name": "Sample Product", "price": 99.99 }
                                        }
                                    ],
                                    "pagination": { "page": 1, "limit": 10, "total": 1, "totalPages": 1, "hasNext": false, "hasPrev": false }
                                }
                            }
                        }
                    }
                },
                "summary": "Get all reviews with advanced filtering",
                "tags": ["Reviews"]
            }
        },
        "/reviews/{id}": {
            "get": {
                "description": "Retrieve a specific review by its ID",
                "operationId": "findOne",
                "parameters": [
                    { "name": "id", "required": true, "in": "path", "description": "Review ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "Review found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "user_id": 1,
                                    "product_id": 1,
                                    "rating": 5,
                                    "title": "Great product!",
                                    "comment": "This product exceeded my expectations.",
                                    "is_verified_purchase": true,
                                    "helpful_votes": 5,
                                    "created_at": "2024-01-01T00:00:00Z",
                                    "updated_at": "2024-01-01T00:00:00Z"
                                }
                            }
                        }
                    },
                    "404": { "description": "Review not found" }
                },
                "summary": "Get review by ID",
                "tags": ["Reviews"]
            },
            "patch": {
                "description": "Update an existing review with new rating, title, or comment. Users can only update their own reviews unless they are admin/moderator.",
                "operationId": "update",
                "parameters": [
                    { "name": "id", "required": true, "in": "path", "description": "Review ID", "schema": { "example": 1, "type": "number" } }
                ],
                "requestBody": {
                    "required": true,
                    "description": "Updated review data",
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateReviewDto" } } }
                },
                "responses": {
                    "200": {
                        "description": "Review updated successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "user_id": 1,
                                    "product_id": 1,
                                    "rating": 4,
                                    "title": "Updated title",
                                    "comment": "Updated comment after longer use.",
                                    "is_verified_purchase": true,
                                    "helpful_votes": 5,
                                    "created_at": "2024-01-01T00:00:00Z",
                                    "updated_at": "2024-01-02T00:00:00Z"
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid input data" },
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot update other user's reviews" },
                    "404": { "description": "Review not found" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Update a review",
                "tags": ["Reviews"]
            },
            "delete": {
                "description": "Delete an existing review by ID. Users can only delete their own reviews unless they are admin/moderator.",
                "operationId": "remove",
                "parameters": [
                    { "name": "id", "required": true, "in": "path", "description": "Review ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted successfully",
                        "content": { "application/json": { "example": { "message": "Review deleted successfully" } } }
                    },
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot delete other user's reviews" },
                    "404": { "description": "Review not found" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Delete a review",
                "tags": ["Reviews"]
            }
        },
        "/reviews/product/{productId}": {
            "get": {
                "description": "Retrieve all reviews for a given product ID",
                "operationId": "findByProduct",
                "parameters": [
                    { "name": "productId", "required": true, "in": "path", "description": "Product ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "Product reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "user_id": 1,
                                        "product_id": 1,
                                        "rating": 5,
                                        "title": "Great product!",
                                        "comment": "This product exceeded my expectations.",
                                        "is_verified_purchase": true,
                                        "helpful_votes": 5,
                                        "created_at": "2024-01-01T00:00:00Z",
                                        "updated_at": "2024-01-01T00:00:00Z"
                                    }
                                ]
                            }
                        }
                    }
                },
                "summary": "Get reviews for a specific product",
                "tags": ["Reviews"]
            }
        },
        "/reviews/product/{productId}/stats": {
            "get": {
                "description": "Get aggregated rating statistics including average rating, total reviews, and rating distribution",
                "operationId": "getProductRatingStats",
                "parameters": [
                    { "name": "productId", "required": true, "in": "path", "description": "Product ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "Rating statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "average_rating": 4.5,
                                    "total_reviews": 10,
                                    "rating_distribution": { "1": 0, "2": 0, "3": 1, "4": 3, "5": 6 }
                                }
                            }
                        }
                    }
                },
                "summary": "Get rating statistics for a product",
                "tags": ["Reviews"]
            }
        },
        "/reviews/user/{userId}": {
            "get": {
                "description": "Retrieve all reviews created by a given user ID. Users can only access their own reviews unless they are admin/moderator.",
                "operationId": "findByUser",
                "parameters": [
                    { "name": "userId", "required": true, "in": "path", "description": "User ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "User reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "user_id": 1,
                                        "product_id": 1,
                                        "rating": 5,
                                        "title": "Great product!",
                                        "comment": "This product exceeded my expectations.",
                                        "is_verified_purchase": true,
                                        "helpful_votes": 5,
                                        "created_at": "2024-01-01T00:00:00Z",
                                        "updated_at": "2024-01-01T00:00:00Z"
                                    }
                                ]
                            }
                        }
                    },
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot access other user's reviews" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Get reviews by a specific user",
                "tags": ["Reviews"]
            }
        },
        "/reviews/{id}/helpful": {
            "patch": {
                "description": "Increment the helpful votes count for a review",
                "operationId": "incrementHelpfulVotes",
                "parameters": [
                    { "name": "id", "required": true, "in": "path", "description": "Review ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "Helpful vote added successfully",
                        "content": {
                            "application/json": { "example": { "id": 1, "helpful_votes": 6, "message": "Helpful vote added successfully" } }
                        }
                    },
                    "404": { "description": "Review not found" }
                },
                "summary": "Mark review as helpful",
                "tags": ["Reviews"]
            }
        }
    },
    "info": {
        "title": "eCommerce API",
        "description": "Comprehensive eCommerce API with advanced querying capabilities including pagination, filtering, sorting, searching, and joins",
        "version": "1.0.0",
        "contact": { "name": "API Support", "url": "https://github.com/yourusername/ecommerce-api", "email": "support@yourcompany.com" },
        "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
    },
...
    "servers": [{ "url": "http://localhost:3151", "description": "Development server" }],
    "components": {
        "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
        "schemas": {
...
            "CreateReviewDto": {
                "type": "object",
                "properties": {
                    "user_id": { "type": "number", "description": "ID of the user creating the review", "example": 1 },
                    "product_id": { "type": "number", "description": "ID of the product being reviewed", "example": 1 },
                    "rating": {
                        "type": "number",
                        "description": "Rating given to the product (1-5 stars)",
                        "example": 5,
                        "minimum": 1,
                        "maximum": 5
                    },
                    "title": { "type": "string", "description": "Optional title for the review", "example": "Great product!", "maxLength": 255 },
                    "comment": {
                        "type": "string",
                        "description": "Detailed comment about the product",
                        "example": "This product exceeded my expectations. The quality is excellent and delivery was fast."
                    },
                    "is_verified_purchase": { "type": "boolean", "description": "Whether this review is from a verified purchase", "example": true }
                },
                "required": ["user_id", "product_id", "rating"]
            },
            "UpdateReviewDto": {
                "type": "object",
                "properties": {
                    "rating": {
                        "type": "number",
                        "description": "Updated rating for the product (1-5 stars)",
                        "example": 4,
                        "minimum": 1,
                        "maximum": 5
                    },
                    "title": { "type": "string", "description": "Updated title for the review", "example": "Updated title", "maxLength": 255 },
                    "comment": {
                        "type": "string",
                        "description": "Updated comment about the product",
                        "example": "Updated my review after using it for a longer time."
                    },
                    "is_verified_purchase": { "type": "boolean", "description": "Updated verified purchase status", "example": false },
                    "helpful_votes": { "type": "number", "description": "Number of helpful votes received", "example": 10, "minimum": 0 }
                }
            }
        }
    }
}